Move App.js one level up to src/ directory and create empty Router file
