+++ PuppyPlayDate/index.android.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native';
-import App from './src/components/App';
+import App from './src/App';
 
 AppRegistry.registerComponent('PuppyPlayDate', () => App);
